Entities added

+++ ATM-BS/ATM-BS.API/Models/BalanceModel.cs
@@ -4,7 +4,7 @@ namespace ATM_BS.API.Models
 {
     public class Balance
     {
-        public int AccountBalance { get; set; }
+        public long AccountBalance { get; set; }
         [ForeignKey("Customer")]
         public int AccountNumber { get; set; }
         public Customer Customer { get; set; }
